chore: move gateway db code to new DbManager struct
